使用Oauth2完成管理员的登录与鉴权，以及会员用户的注册与登录，登录使用jwt，购物车模块，Mq信息模块，图片验证码模块，邮箱消息模块、

+++ provider/provider-goods/src/main/java/com/akali/provider/goods/bean/PmsBaseAttrOption.java
@@ -34,6 +34,6 @@ public class PmsBaseAttrOption {
      * 数值类型的值
      */
     @Column(length=10)
-    private Integer numValue;
+    private Float numValue;
 
 }
